# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Platform API
  version: 1.0.0
  description: Multi-tenant B2B inventory & ordering platform. All money values are
    decimal strings.
paths:
...
  /api/v1/auth/handoff/exchange/:
    post:
      operationId: auth_handoff_exchange
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoffExchangeInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/impersonation/act/:
    post:
      operationId: auth_impersonation_act
      description: Switch the current support session to ACT mode (a reason is required;
        audited).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpersonationReasonRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: ''
  /api/v1/auth/impersonation/end/:
    post:
      operationId: auth_impersonation_end
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/auth/logout/:
    post:
      operationId: auth_logout
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshInputRequest'
      security:
      - {}
      responses:
        '204':
          description: No response body
  /api/v1/auth/me/:
    get:
      operationId: auth_me_retrieve
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
    patch:
      operationId: auth_me_update
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMeUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
  /api/v1/auth/mfa/confirm/:
    post:
      operationId: auth_mfa_confirm
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaSetupConfirmInputRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodes'
          description: ''
  /api/v1/auth/mfa/disable/:
    post:
      operationId: auth_mfa_disable
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordAndFactorInputRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/auth/mfa/recovery-codes/:
    post:
      operationId: auth_mfa_recovery_codes
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordAndFactorInputRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodes'
          description: ''
  /api/v1/auth/mfa/setup/:
    post:
      operationId: auth_mfa_setup
      tags:
      - auth
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSetupResponse'
          description: ''
  /api/v1/auth/password/change/:
    post:
      operationId: auth_password_change
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeInputRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: ''
  /api/v1/auth/password/forgot/:
    post:
      operationId: auth_password_forgot
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordForgotInputRequest'
        required: true
      security:
      - {}
      responses:
        '202':
          description: No response body
  /api/v1/auth/password/reset/:
    post:
      operationId: auth_password_reset
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetInputRequest'
        required: true
      security:
      - {}
      responses:
        '204':
          description: No response body
  /api/v1/auth/retailer/choose-account/:
    post:
      operationId: auth_retailer_choose_account
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetailerChooseAccountInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/retailer/otp/request/:
    post:
      operationId: auth_retailer_otp_request
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetailerOtpRequestInputRequest'
        required: true
      security:
      - {}
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetailerOtpRequestResponse'
          description: ''
  /api/v1/auth/retailer/otp/verify/:
    post:
      operationId: auth_retailer_otp_verify
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetailerOtpVerifyInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/staff/choose-tenant/:
    post:
      operationId: auth_staff_choose_tenant
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChooseTenantInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/staff/login/:
    post:
      operationId: auth_staff_login
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffLoginInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/staff/mfa/enrol/confirm/:
    post:
      operationId: auth_staff_mfa_enrol_confirm
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrolmentConfirmInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/staff/mfa/enrol/start/:
    post:
      operationId: auth_staff_mfa_enrol_start
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrolmentTokenInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSecret'
          description: ''
  /api/v1/auth/staff/mfa/verify/:
    post:
      operationId: auth_staff_mfa_verify
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyInputRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
  /api/v1/auth/token/refresh/:
    post:
      operationId: auth_token_refresh
      description: No authentication, any caller; and no request-wide transaction
        (see module docstring).
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshInputRequest'
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: ''
...
components:
  schemas:
...
    HandoffExchangeInputRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 200
      required:
      - code
...
    ImpersonationReasonRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
      required:
      - reason
...
    LoginResponse:
      type: object
      description: One shape for every sign-in step; ``status`` says what the client
        does next.
      properties:
        status:
          $ref: '#/components/schemas/LoginStatusEnum'
        user_type:
          $ref: '#/components/schemas/UserTypeEnum'
        access:
          type: string
        access_expires_at:
          type: string
          format: date-time
        handoff:
          $ref: '#/components/schemas/Handoff'
        choice_token:
          type: string
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/TenantChoice'
        mfa_token:
          type: string
        enrolment_token:
          type: string
        recovery_codes:
          type: array
          items:
            type: string
          description: Shown once, right after set-up.
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/RetailerAccountChoice'
      required:
      - status
...
    Me:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_type:
          $ref: '#/components/schemas/UserTypeEnum'
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        full_name:
          type: string
        preferred_language:
          type: string
        tenant:
          allOf:
          - $ref: '#/components/schemas/MeTenant'
          nullable: true
        role:
          allOf:
          - $ref: '#/components/schemas/MeRole'
          nullable: true
        retailer:
          allOf:
          - $ref: '#/components/schemas/MeRetailer'
          nullable: true
        permissions:
          type: array
          items:
            type: string
        features:
          type: object
          additionalProperties:
            type: boolean
        impersonation:
          allOf:
          - $ref: '#/components/schemas/MeImpersonation'
          nullable: true
        mfa_enabled:
          type: boolean
        mfa_required:
          type: boolean
          description: 2FA can't be turned off for this account.
      required:
      - email
      - features
      - full_name
      - id
      - impersonation
      - mfa_enabled
      - mfa_required
      - permissions
      - phone
      - preferred_language
      - retailer
      - role
      - tenant
      - user_type
...
    RefreshInputRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
          maxLength: 2000
...
    TokenResponse:
      type: object
      properties:
        access:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh:
          type: string
          description: Only when the request sent one.
      required:
      - access
      - access_expires_at
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT